requirements.txt
# SMYLE PLAY — Dépendances Python
# Installation : pip3 install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────────
flask>=3.0.0
gunicorn>=21.2.0

# ── Base de données ───────────────────────────────────────────────────────────
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.9       # PostgreSQL (Railway / production)

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── Cloudflare R2 / AWS S3 ────────────────────────────────────────────────────
boto3>=1.34.0

# ── Sécurité (inclus avec Flask, listé pour clarté) ──────────────────────────
# werkzeug>=3.0.0              # hash passwords — déjà inclus dans flask
